models/HIPTrack/requirements.txt
# Python packages from enviroment.yaml
# Note: Python version should be 3.8 (use: uv venv --python 3.8)
# 
# Installation instructions:
# 1. uv venv --python 3.8
# 2. source .venv/bin/activate
# 3. uv pip install torch==1.10.1+cu113 torchvision==0.11.2+cu113 torchaudio==0.10.1 --index-url https://download.pytorch.org/whl/cu113
# 4. uv pip install cython numpy setuptools wheel
# 5. uv pip install -r requirements.txt --no-build-isolation

# Build dependencies (install first)
cython>=0.29.30
numpy>=1.22.3
setuptools
wheel

# Core dependencies (PyTorch already installed in step 3)
pillow==9.1.1
matplotlib==3.5.2
jupyter==1.0.0
ipykernel==6.13.0
ipython==8.4.0
notebook==6.4.11
jupyterlab-widgets==1.1.0
ipywidgets==7.7.0
qtconsole==5.3.0

# Pip dependencies (from original environment.yaml)
absl-py==1.1.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astroid==2.12.5
attributee==0.1.8
audioread==3.0.1
auraloss==0.4.0
autoflake==2.0.1
basecore==0.1.9
bidict==0.22.1
black==21.4b2
boto3==1.24.42
botocore==1.27.42
cachetools==5.2.0
cityscapesscripts==2.2.0
click==8.1.3
clip==0.2.0
cloudpickle==2.1.0
colorama==0.4.4
coloredlogs==15.0.1
commonmark==0.9.1
crc32c==2.3
crcmod==1.7
cycler==0.11.0
cython==0.29.30
cython-bbox==0.1.3
decorator==4.4.2
diffdist==0.1
dill==0.3.5.1
docker-pycreds==0.4.0
dominate==2.8.0
easydict==1.9
einops==0.7.0
et-xmlfile==1.1.0
evdev==1.6.0
fairscale==0.4.6
faiss-cpu==1.7.2
filterpy==1.4.5
flatbuffers==2.0
fonttools==4.33.3
fsspec==2023.10.0
ftfy==6.1.3
future==0.18.2
fvcore==0.1.5.post20220512
gevent==23.9.1
gitdb==4.0.9
gitpython==3.1.27
# google-auth==2.23.4
# google-auth-oauthlib==1.0.0
greenlet==3.0.1
grpcio==1.59.2
h5py==3.7.0
huggingface-hub==0.19.4
humanfriendly==10.0
hydra-core==1.2.0
imageio==2.19.3
imageio-ffmpeg==0.4.7
iopath==0.1.9
isort==5.10.1
jmespath==0.10.0
joblib==1.1.0
jpeg4py==0.1.4
jsonpatch==1.32
jsonpointer==2.2
kiwisolver==1.4.2
# lap==0.4.0  # Already installed separately
lazy-object-proxy==1.9.0
librosa==0.8.1
llvmlite==0.39.0
lmdb==1.3.0
loguru==0.6.0
markdown==3.3.7
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
# megengine==1.11.0+cu114  # Not available on PyPI with CUDA tag, install separately if needed
memory-profiler==0.61.0
mmcv-full==1.7.1
mmdet==2.28.1
model-index==0.1.11
motmetrics==1.2.5
moviepy==0.2.3.5
mpmath==1.2.1
# multiscaledeformableattention==1.0  # Not available on PyPI, may need to install from source
mypy-extensions==0.4.3
natsort==8.1.0
networkx==2.8.2
ninja==1.10.2.3
numba==0.56.0
oauthlib==3.2.0
omegaconf==2.2.2
onnx==1.12.0
onnx-simplifier==0.4.5
onnxoptimizer==0.2.7
onnxruntime==1.12.1
# onnxruntime-gpu==1.6.0  # Conflict with onnxruntime, use onnxruntime-gpu instead if GPU needed
opencv-python==4.5.5.64
openmim==0.3.5
openpyxl==3.0.10
openxlab==0.0.33
ordered-set==4.1.0
oss2==2.17.0
pandas==1.4.2
# panopticapi==0.1  # Not available on PyPI, install from git: pip install git+https://github.com/cocodataset/panopticapi.git
pathspec==0.9.0
pathtools==0.1.2
pesq==0.0.4
phx-class-registry==4.0.6
platformdirs==2.5.2
pooch==1.8.0
portalocker==2.5.1
prettytable==3.3.0
progressbar==2.5
promise==2.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycryptodome==3.19.1
pydot==1.4.2
pyflakes==3.0.1
pygments==2.17.2
pylatex==1.4.2
pylint==2.15.0
pynput==1.7.6
pyparsing==3.1.1
pyqt5==5.15.7
pyqt5-qt5==5.15.2
pyqt5-sip==12.11.0
pyquaternion==0.9.9
python-xlib==0.31
pytz==2023.3.post1
pywavelets==1.3.0
pyyaml==6.0
regex==2022.8.17
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.4.2
rsa==4.8
s3transfer==0.6.0
safetensors==0.4.0
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
# script-utils==0.0.1  # Not available on PyPI
seaborn==0.11.2
sentry-sdk==1.6.0
setproctitle==1.2.3
shapely==1.7.0
shortuuid==1.0.9
smmap==5.0.0
soundfile==0.12.1
submitit==1.4.2
sympy==1.10.1
tabulate==0.8.9
# tao==0.1.0  # Not available on PyPI
# tb-nightly==2.14.0a20230808  # Conflicts with tensorboard
tensorboard==2.9.0
# tensorboard-data-server==0.7.2  # Version conflict, let tensorboard install compatible version
tensorboard-plugin-wit==1.8.1
tensorboardx==2.5.1
termcolor==1.1.0
terminaltables==3.1.10
# thinplate==1.0.0
# thop==0.1.1.post2311151112
threadpoolctl==3.1.0
tifffile==2022.5.4
tikzplotlib==0.10.1
timm==0.5.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
torch-summary==1.4.5
torchfile==0.1.0
torchstat==0.0.7
tqdm==4.65.2
typing==3.7.4.3
tzdata==2023.3
visdom==0.1.8.9
vot-toolkit==0.6.4
vot-trax==4.0.1
wandb==0.12.20
wcwidth==0.2.13
webcolors==1.12
websocket==0.2.1
websocket-client==1.3.2
werkzeug==2.1.2
wget==3.2
wrapt==1.14.1
xmltodict==0.13.0
yacs==0.1.8
yapf==0.32.0
youtube-dl==2021.12.17
zope-event==5.0
zope-interface==6.1

